Main: fix units

+++ include/kernaux/units.h
@@ -7,6 +7,7 @@ extern "C" {
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 enum KernAux_Unit {
     KERNAUX_UNIT_BIT,
@@ -25,12 +26,12 @@ enum KernAux_UnitPrefixBin {
     KERNAUX_UNITPFX_GIBI,
 };
 
-bool kernaux_units_human_raw(size_t value, enum KernAux_Unit unit,
+bool kernaux_units_human_raw(uint64_t value, enum KernAux_Unit unit,
                              char *buffer, size_t buffer_size);
-bool kernaux_units_human_dec(size_t value, enum KernAux_Unit unit,
+bool kernaux_units_human_dec(uint64_t value, enum KernAux_Unit unit,
                              enum KernAux_UnitPrefixDec prefix,
                              char *buffer, size_t buffer_size);
-bool kernaux_units_human_bin(size_t value, enum KernAux_Unit unit,
+bool kernaux_units_human_bin(uint64_t value, enum KernAux_Unit unit,
                              enum KernAux_UnitPrefixBin prefix,
                              char *buffer, size_t buffer_size);
